refactor(queries): split query engine into smaller files
Split the two largest query engine files into focused, single-concern
modules to improve maintainability and prepare for upcoming fixes.

query.c (1156 lines) split into:
- query.c (408) - entry points: rrd2rrdr(), rrd2rrdr_legacy()
- query-execute.c (451) - dimension query execution loop, interpolation
- query-window.c (303) - query window calculation

query-group-by.c (1383 lines) split into:
- query-group-by.c (226) - parsing, JSON helpers, label utilities
- query-group-by-init.c (516) - group-by initialization
- query-group-by-finalize.c (411) - finalization, trimming, aggregation
- query-cardinality-limit.c (235) - cardinality limiting

Zero behavior changes. All files compile clean with no duplicate symbols.
